# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/repair_parts:
    get:
      tags:
        - RepairParts
      summary: List RepairParts
      description: Get a paginated list of RepairParts with optional filtering and sorting
      operationId: listRepairParts
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepairPartList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - RepairParts
      summary: Create RepairPart
      description: Create a new RepairPart
      operationId: createRepairPart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRepairPartRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepairPart'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/repair_tags:
    get:
      tags:
        - RepairTags
      summary: List RepairTags
      description: Get a paginated list of RepairTags with optional filtering and sorting
      operationId: listRepairTags
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepairTagList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - RepairTags
      summary: Create RepairTag
      description: Create a new RepairTag
      operationId: createRepairTag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRepairTagRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepairTag'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    RepairPart:
      type: object
      required:
        - id
        - company_id
        - repair_order_id
        - item_id
        - line_type
        - quantity
        - rate
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        repair_order_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        warehouse_id:
          nullable: true
          type: string
          format: uuid
        line_type:
          type: string
        quantity:
          type: string
        rate:
          type: string
        metadata:
          type: object
...
    CreateRepairPartRequest:
      type: object
      required:
        - company_id
        - repair_order_id
        - item_id
        - line_type
        - quantity
        - rate
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        repair_order_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        warehouse_id:
          nullable: true
          type: string
          format: uuid
        line_type:
          type: string
        quantity:
          type: string
        rate:
          type: string
        metadata:
          type: object
...
    RepairPartList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RepairPart'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    RepairTag:
      type: object
      required:
        - id
        - company_id
        - name
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 140
        metadata:
          type: object
...
    CreateRepairTagRequest:
      type: object
      required:
        - company_id
        - name
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 140
        metadata:
          type: object
...
    RepairTagList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RepairTag'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []